+++ test/components/notifications/Notifications.spec.js
@@ -5,8 +5,7 @@ import sinon from 'sinon';
 
 import { Notifications } from '~/components/notifications/Notifications';
 
-import { api, fakeAPI } from '@/data';
-import { seenEvent } from '@/data/events';
+import { api } from '@/data';
 
 const events = api.events;
 
@@ -46,44 +45,6 @@ describe('components/Notifications', function () {
     expect(notifications.instance().fetchAllEvents).to.exist;
 
     notifications.instance().fetchAllEvents();
-
-    expect(dispatch.callCount).to.equal(1);
-  });
-
-  it('implements making events seen', () => {
-    const emptyEventsAPI = fakeAPI([
-      [{}, 'event', 'events'],
-    ]);
-    notifications = shallow(<Notifications dispatch={dispatch} events={emptyEventsAPI.events} />);
-    expect(notifications.instance().markEventsSeen).to.exist;
-
-    notifications.instance().markEventsSeen();
-
-    expect(dispatch.callCount).to.equal(0);
-  });
-
-  it('only marks unseen events seen', () => {
-    const seenEventsAPI = fakeAPI([
-      [
-        {
-          [seenEvent.id]: seenEvent,
-        },
-        'event',
-        'events',
-      ],
-    ]);
-    notifications = shallow(<Notifications dispatch={dispatch} events={seenEventsAPI.events} />);
-    expect(notifications.instance().markEventsSeen).to.exist;
-
-    notifications.instance().markEventsSeen();
-
-    expect(dispatch.callCount).to.equal(0);
-  });
-
-  it('marks events seen for all unseen events', () => {
-    expect(notifications.instance().markEventsSeen).to.exist;
-
-    notifications.instance().markEventsSeen();
 
     expect(dispatch.callCount).to.equal(1);
   });
